chore: Moves modal dialog files to components
